+++ R/VCF_to_catalog_functions.R
@@ -646,7 +646,14 @@ ReadVCF <-
     if (nrow(df0) == 0) {
       return(df0)
     }
-
+    
+    # Check whether df0 has column name "FILTER"
+    if (!"FILTER" %in% colnames(df0)) {
+      warning("\nThere is no column FILTER in the file ", file,
+              "\nAll variants will be retained")
+      filter.status <- NULL
+    }
+    
     # Remove rows that don't have the specified filter status
     if (is.null(filter.status)) {
       df1 <- df <- df0